requirements.txt
bitsandbytes==0.45.3
deepspeed==0.16.7
editdistance==0.8.1
fastapi==0.115.12
flash_attn==2.7.4.post1
gradio==5.29.0
huggingface_hub==0.27.0
matplotlib==3.10.1
numpy==1.25.1
openai==1.77.0
packaging==25.0
pandas==2.2.3
peft==0.15.0
Pillow==11.2.1
pycocoevalcap==1.2
ray==2.40.0
Requests==2.32.3
scipy==1.15.2
seaborn==0.13.2
sglang==0.4.6.post2
shortuuid==1.0.13
tokenizers==0.21.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.2
uvicorn==0.34.2
xformers==0.0.28.post3
